vision/pyproject.toml
[tool.poetry]
name = "vision"
version = "0.1.0"
description = "Computer vision for 2036"
authors = ["Fairview Robotics <fairviewrobotics@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.23.2"
opencv-python = "^4.6.0"
pynetworktables = "^2021.0.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
